tests for z, include index, empty/none
